Module sw8s_rust_graphs::generated_actions::action

Structs

ActionChain

ActionConcurrent

ActionConcurrentSplit

ActionConditional

An action that runs one of two actions depending on if its conditional reference is true or false.

ActionDataConditional

An action that runs one of two actions depending on if its conditionalreference is valid or not, passing processed output to the first child andunprocessed output to the second child.

ActionParallel

ActionSequence

ActionUntil

An action that tries count times for a success

ActionWhile

An action that runs while true

DualAction

Run two actions at once, and only exit when all actions have exited.

FirstValid

Return first valid response from block of actions.

RaceAction

Action that runs two actions at the same time and exits both when one exits

TupleSecond

Get second arg in action output

Traits

Action

A trait for an action that can be executed.Functions returning Actions/ActionExec should be written with -> impl ActionExec + '_ andusing Con for the context generic. This allows the build script to strip out the context andcreate a mirrored version returning -> impl Action +'_ for graphing.

ActionExec

A trait for an action that can be executed.

ActionIgnoredGeneric

ActionMod

An action that can be modified at runtime.

Functions

graph_actions

wrap_action

Produces a new function that puts wrapper around wrapee.Needed for act_nest! macros that want to combine MetaAction::new calls,since the recursive construction doesn’t use identical values at each call site.e.g. wrap_action(ActionConcurrent::new, FirstValid::new) in order to run thenested actions concurrently, outputting the first valid return