Module sw8s_rust_graphs::generated_actions::movement

Structs

AdjustMovement

AdjustMovementAngle

BoxToPose

CautiousConstantX

CenterMovement

ClampX

ConfidenceY

ConstYaw

DefaultGen

Descend

FlatX

FlipX

FlipYaw

GlobalMovement

GlobalPos

Stores the command to send to stability assist 2

InvertX

LinearYawFromX

Action version of linear_yaw_from_x

MinYaw

MultiplyX

OffsetToPose

ReplaceX

SetSideBlue

SetSideRed

SetX

SetY

SideIsRed

SideMult

Stability1Adjust

Modification for a stability assist 1 command

Stability1Movement

Stability1Pos

Stores the command to send to stability assist 2

Stability2Adjust

Modification for a stability assist 2 command

Stability2Movement

Stability2Pos

Stores the command to send to stability assist 2

StraightMovement

StripX

StripY

ZeroMovement

Enums

AdjustType

Specifies replacement or adjustment (+ value)

Side

Statics

SIDE 🔒

Functions

aggressive_yaw_from_x

default_linear_yaw_from_x

linear_yaw_from_x with a default value

graph_actions

linear_yaw_from_x

Generates a yaw adjustment from an x axis set, multiplying by angle_diff

linear_yaw_from_x_stab1

Generates a yaw adjustment from an x axis set, multiplying by angle_diff