sw8s_rust_graphs::generated_actions
Module vision

Structs

Average

DetectTarget

ExtractPosition

MidPoint

Norm

Normalizes vision output.

OffsetClass

SizeUnder

ToOffset

Vision

Runs a vision routine to obtain object positions.

VisionNorm

Runs a vision routine to obtain object positions

VisionNormBottom

Runs a vision routine to obtain object positions

VisionNormOffset

Runs a vision routine to obtain the average of object positions

VisionNormOffsetBottom

Runs a vision routine to obtain the average of object positions

VisionSizeLock

Runs a vision routine to obtain object positions.

Statics

PIPELINE_KILL

Functions

graph_actions