Module sw8s_rust_graphs::generated_actions::movement

Structs

AdjustMovement

AdjustMovementAngle

CenterMovement

Descend

FlatX

LinearYawFromX

Action version of linear_yaw_from_x

OffsetToPose

Stability1Adjust

Modification for a stability assist 1 command

Stability1Movement

Stability1Pos

Stores the command to send to stability assist 2

Stability2Adjust

Modification for a stability assist 2 command

Stability2Movement

Stability2Pos

Stores the command to send to stability assist 2

StraightMovement

StripX

StripY

ZeroMovement

Enums

AdjustType

Specifies replacement or adjustment (+ value)

Functions

aggressive_yaw_from_x

default_linear_yaw_from_x

linear_yaw_from_x with a default value

graph_actions

linear_yaw_from_x

Generates a yaw adjustment from an x axis set, multiplying by angle_diff